Handle OFX emitter which closes text tags
Typically the STMTTRN lines look something like this:
  <STMTTRN><TRNTYPE>XFER<DTPOSTED>20170821120000.000[-5:EST]

Some vendors (like HomeTrust via EZCard) emits lines which look like this:
  <STMTTRN><TRNTYPE>XFER</TRNTYPE><DTPOSTED>20170821120000.000[-5:EST]</DTPOSTED>

OFX::Tree was not looking for closing of the current tag and was blindly closing it once for the text and a second time for the actual end tag. Skip the second closing for this type of file.